........
  r708385 | josh | 2008-10-27 16:44:44 -0700 (Mon, 27 Oct 2008) | 1 line
  
  Small fix for bug in RecordInputStream.readAllContinuedRemainder()  introduced in r707778.  It seems like only BiffViewer was affected.
........
  r708996 | josh | 2008-10-29 13:13:58 -0700 (Wed, 29 Oct 2008) | 1 line
  
  Allowed for quad-byte padding alignment on ObjRecord
........
  r709054 | josh | 2008-10-29 17:21:24 -0700 (Wed, 29 Oct 2008) | 1 line
  
  removed obsolete methods
........
  r709217 | josh | 2008-10-30 10:56:34 -0700 (Thu, 30 Oct 2008) | 1 line
  
  Fixed compiler warnings / simplified code
........
  r709221 | josh | 2008-10-30 11:33:35 -0700 (Thu, 30 Oct 2008) | 1 line
  
  Optimised slow test case (after reviewing original purpose)
........
  r709235 | josh | 2008-10-30 13:17:04 -0700 (Thu, 30 Oct 2008) | 1 line
  
  Fix for bug 15716 - - converted Ptg arrays into Formula objects to optimise memory usage
........


git-svn-id: https://svn.apache.org/repos/asf/poi/branches/ooxml@709262 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hssf/usermodel/HSSFPatternFormatting.java
@@ -87,7 +87,7 @@ public class HSSFPatternFormatting
 	 */
 	public short getFillBackgroundColor()
 	{
-		return patternFormatting.getFillBackgroundColor();
+		return (short)patternFormatting.getFillBackgroundColor();
 	}
 
 	/**
@@ -96,7 +96,7 @@ public class HSSFPatternFormatting
 	 */
 	public short getFillForegroundColor()
 	{
-		return patternFormatting.getFillForegroundColor();
+		return (short)patternFormatting.getFillForegroundColor();
 	}
 
 	/**
@@ -105,7 +105,7 @@ public class HSSFPatternFormatting
 	 */
 	public short getFillPattern()
 	{
-		return patternFormatting.getFillPattern();
+		return (short)patternFormatting.getFillPattern();
 	}
 
 	/**
